+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Improvements:
 * [OLMIS-3955](https://openlmis.atlassian.net/browse/OLMIS-3955): Renamed PICKING order status to FULFILLING. Removed PICKED and IN_TRANSIT.
 * [OLMIS-3925](https://openlmis.atlassian.net/browse/OLMIS-3925): Updated requisition approval process to stop at an intermediate supervisory node that has a supply line for the requisition's program.
 * [OLMIS-3930](https://openlmis.atlassian.net/browse/OLMIS-3930): a requisition will be converted to an order if for a supplying facility/program combo locally fulfill flags is set
+* [OLMIS-3938](https://openlmis.atlassian.net/browse/OLMIS-3938): Configured ISA column on requisition template.
 
 5.1.0 / 2017-11-09
 ==================